test: rename and split VFS test files into topic-based names
Replaces the -coverage / -branches / -misc suffixes with focused
files named after the API or behaviour they exercise. Splits the
larger multi-topic files into one-topic-per-file.

Renames:
- callbacks.js → callback-api.js
- stats-defaults.js → stats-helpers.js
- file-handle-base.js → virtual-file-handle.js
- provider-base.js → virtual-provider.js
- provider-memory.js → memory-provider.js
- real-provider-async.js → real-provider-promises.js
- mkdir-recursive-return.js → mkdir.js

New files (split out of -coverage/-branches/-misc):
- access-modes, create, link, mkdtemp, rename, symlinks, utimes,
  write-options
- memory-file-handle, memory-provider-dynamic, memory-provider-flags
- real-provider-handle, real-provider-symlinks, real-provider-watch
- stream-errors, stream-explicit-fd
- watch, watch-abort-signal, watch-encoding, watch-promises,
  watch-recursive

Coverage maintained at 97.6% line / 95.2% branch / 95.3% function.
