+++ test/parallel/test-vfs-virtual-file-handle.js
