+++ test/parallel/test-vfs-virtual-provider.js
