+++ test/parallel/test-vfs-watch-promises.js
@@ -1,0 +1,65 @@
+'use strict';
+
+// promises.watch() returns an async iterable. Cover its event queue,
+// next/return/throw, and close-while-pending behaviour.
+
+const common = require('../common');
+const assert = require('assert');
+const vfs = require('node:vfs');
+
+(async () => {
+  // Basic for-await iteration receives a change event
+  {
+    const myVfs = vfs.create();
+    myVfs.writeFileSync('/file.txt', 'a');
+    const iter = myVfs.promises.watch('/file.txt', { interval: 25 });
+    queueMicrotask(() => myVfs.writeFileSync('/file.txt', 'longer-changed'));
+    for await (const evt of iter) {
+      assert.strictEqual(evt.eventType, 'change');
+      break;
+    }
+  }
+
+  // Events queued before next() drain via the next call
+  {
+    const myVfs = vfs.create();
+    myVfs.writeFileSync('/q.txt', 'a');
+    const iter = myVfs.promises.watch('/q.txt', { interval: 25 });
+    myVfs.writeFileSync('/q.txt', 'longer-changed');
+    const r = await iter.next();
+    if (!r.done) assert.strictEqual(r.value.eventType, 'change');
+    await iter.return();
+  }
+
+  // A change while a next() is pending shifts the resolver
+  {
+    const myVfs = vfs.create();
+    myVfs.writeFileSync('/q2.txt', 'a');
+    const iter = myVfs.promises.watch('/q2.txt', { interval: 25 });
+    const pending = iter.next();
+    myVfs.writeFileSync('/q2.txt', 'longer-changed');
+    const r = await pending;
+    if (!r.done) assert.strictEqual(r.value.eventType, 'change');
+    await iter.return();
+  }
+
+  // throw() closes the watcher and resolves with done:true
+  {
+    const myVfs = vfs.create();
+    myVfs.writeFileSync('/q3.txt', 'a');
+    const iter = myVfs.promises.watch('/q3.txt', { interval: 1000 });
+    const r = await iter.throw(new Error('go away'));
+    assert.strictEqual(r.done, true);
+  }
+
+  // close while a resolver is pending — drains via the 'close' handler
+  {
+    const myVfs = vfs.create();
+    myVfs.writeFileSync('/q4.txt', 'a');
+    const iter = myVfs.promises.watch('/q4.txt', { interval: 1000 });
+    const pending = iter.next();
+    queueMicrotask(() => iter.return());
+    const r = await pending;
+    assert.strictEqual(r.done, true);
+  }
+})().then(common.mustCall());
